backend/pyproject.toml
[project]
name = "proofbot-backend"
version = "0.1.0"
description = "Telegram bot and TWA backend for startup idea analysis MVP"
requires-python = ">=3.12"
dependencies = [
  "aiofiles>=24.1.0",
  "aiogram>=3.18.0",
  "alembic>=1.14.1",
  "arq>=0.26.3",
  "asyncpg>=0.30.0",
  "fastapi>=0.115.8",
  "greenlet>=3.1.1",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "jinja2>=3.1.5",
  "openai>=1.66.3",
  "orjson>=3.10.15",
  "pydantic-settings>=2.8.1",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.38",
  "structlog>=25.1.0",
  "tenacity>=9.0.0",
  "uvicorn[standard]>=0.34.0",
  "weasyprint>=64.1"
]

[project.optional-dependencies]
dev = [
  "aiosqlite>=0.21.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "respx>=0.22.0",
]

[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
